requirements.txt
pytest~=8.4.1
allure-pytest
selene>=2.0.b4
allure-python-commons~=2.14.3
webdriver-manager
selene
selenium~=4.34.0
python-dotenv~=1.1.1